Use Mach time  when clock_gettime() not suppported
Older versions of MacOS did not support POSIX `clock_gettime()`, but newer do.
So, we have a runtime check to see if we're running on a MacOS that does support
it, and use the Mach kernel time APIs for older MacOSes. This does incur a runtime
check cost, but is a clear path forward for when the support for older MacOSes is
dropped.