Use gettimeofday() on Mac
Using clock_gettime() with `CLOCK_REALTIME` is actually the same as using gettimeofday() - as per the Apple docs, yet clock_gettime() is not available on older MacOS versions. So this change in config doesn't change the behaviour, it only lets us run on older Macs.

This is a short-term fix, in the long run, we should actually use a monotonic clock, clock_gettime() for newer OSes and Mach kernel APIs for older.